Update stage1.cpp
built out a couple more of the productions, mainly working on getting the structure in so that we can just run through all the code files before we have to get it doing the ops, lets us work on these separately from the other stuff below it will need in the future.
Very unsure about how i'm handling express, my idea is that it basically is just there to call other productions that check it and keep everything structured. Its a shit ton of recursion though so it get tricky tracking everything. I added a couple dummy returns and got everything to compile so we can do it piece by piece now at least making sure there arent syntax errors.